Moved ErrorResponse to responses folder
